Catch/Release and 1st statistic

+++ Application/DTOs/Pokemon/PokemonDto.cs
@@ -5,5 +5,5 @@ public class PokemonDto : PokemonManipulationDto, IEntityDto
 {
     public int Id { get; set; }
     public IEnumerable<PowerDto> Powers { get; set; } = [];
-    public ulong CatchCount { get; set; }
+    public double CatchCount { get; set; }
 }
